requirements.txt
aiogram~=2.8
aiohttp~=3.6.2
aiosqlite~=0.13.0
asyncpg~=0.20.1
PyPika~=0.37.7
pytz~=2020.1
sentry-sdk~=0.14.4
tortoise-orm~=0.16.12
aiodns==2.0.0
cchardet==2.1.6
ujson==2.0.3
uvloop==0.14.0
